# swagger.yml
openapi: 3.0.2
info:
  title: Autograder.io API
  version: 3.5.2
paths:
...
  /api/projects/{id}/:
    get:
      operationId: RetrieveProjectDetail
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
      tags: &id015
      - projects
    patch:
      operationId: PartialUpdateProjectDetail
      description: ''
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: ''
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pk: *id008
                name:
                  description: "The name used to identify this project.\n        \
                    \    Must be non-empty and non-null.\n            Must be unique\
                    \ among Projects associated with\n            a given course.\n\
                    \            This field is REQUIRED."
                  nullable: false
                  readOnly: false
                  type: string
                last_modified:
                  description: ''
                  nullable: false
                  readOnly: true
                  type: string
                  format: date-time
                course:
                  description: "The Course this project belongs to.\n            This\
                    \ field is REQUIRED."
                  nullable: false
                  readOnly: true
                  type: integer
                  format: id
                visible_to_students:
                  description: "Whether information about this Project can\n     \
                    \       be viewed by students."
                  nullable: false
                  readOnly: false
                  type: boolean
                closing_time:
                  description: "The date and time that this project should stop\n\
                    \            accepting submissions.\n            A value of None\
                    \ indicates that this project should\n            stay open.\n\
                    \            In the API, this field is hidden from non-admins."
                  nullable: true
                  readOnly: false
                  type: string
                  format: date-time
                soft_closing_time:
                  description: "The date and time that should be displayed as the\n\
                    \            due date for this project. Unlike closing_time,\n\
                    \            soft_closing_time does not affect whether submissions\
                    \ are\n            actually accepted.\n            If not None\
                    \ and closing_time is not None, this value must be\n         \
                    \   less than (before) closing_time."
                  nullable: true
                  readOnly: false
                  type: string
                  format: date-time
                disallow_student_submissions:
                  description: "A hard override that indicates that students should\n\
                    \            be prevented from submitting even if visible_to_students\
                    \ is\n            True and it is before closing_time."
                  nullable: false
                  readOnly: false
                  type: boolean
                disallow_group_registration:
                  description: "A hard override that indicates that students should\n\
                    \            not be able to send, accept, or reject group\n  \
                    \          invitations."
                  nullable: false
                  readOnly: false
                  type: boolean
                guests_can_submit:
                  description: "By default, only admins, staff, and students\n   \
                    \         for a given Course can view and submit to its Projects.\n\
                    \            When True, submissions will be accepted from guests\n\
                    \            with the following caveats:\n                - Guests\
                    \ must be given a direct link to the project.\n              \
                    \  - When group work is allowed, guests can\n                only\
                    \ be in groups with other guests."
                  nullable: false
                  readOnly: false
                  type: boolean
                min_group_size:
                  description: "The minimum number of students that can work in a\n\
                    \            group on this project.\n            Must be >= 1.\n\
                    \            Must be <= max_group_size."
                  nullable: false
                  readOnly: false
                  type: integer
                max_group_size:
                  description: "The maximum number of students that can work in a\n\
                    \            group on this project.\n            Must be >= 1.\n\
                    \            Must be >= min_group_size."
                  nullable: false
                  readOnly: false
                  type: integer
                submission_limit_per_day:
                  description: "The number of submissions each group is allowed per\n\
                    \            day before either reducing feedback or preventing\
                    \ further\n            submissions. A value of None indicates\
                    \ no limit."
                  nullable: true
                  readOnly: false
                  type: integer
                allow_submissions_past_limit:
                  description: "Whether to allow additional submissions after a\n\
                    \            group has submitted submission_limit_per_day times."
                  nullable: false
                  readOnly: false
                  type: boolean
                groups_combine_daily_submissions:
                  description: "If True, group members can \"pool\" their daily submissions.\n\
                    \            For example, if submission_limit_per_day is 3,\n\
                    \            a group with 2 members would get 6 submissions per\
                    \ day."
                  nullable: false
                  readOnly: false
                  type: boolean
                submission_limit_reset_time:
                  description: "The time that marks the beginning and end of the 24\n\
                    \            hour period during which submissions should be counted\n\
                    \            towards the daily limit. Defaults to 00:00:00."
                  nullable: false
                  readOnly: false
                  type: string
                  format: time
                submission_limit_reset_timezone:
                  description: "The timezone to use when computing how many\n    \
                    \        submissions a group has made in a 24 hour period."
                  nullable: false
                  readOnly: false
                  type: string
                  format: timezone
                num_bonus_submissions:
                  description: ''
                  nullable: false
                  readOnly: false
                  type: integer
                total_submission_limit:
                  description: "The maximum number of times a Group can submit to\n\
                    \            this Project EVER."
                  nullable: true
                  readOnly: false
                  type: integer
                allow_late_days:
                  description: "Whether to allow the use of late days for submitting\n\
                    \            past the deadline."
                  nullable: false
                  readOnly: false
                  type: boolean
                ultimate_submission_policy:
                  description: "The \"ultimate\" submission for a group is the one\n\
                    \            that will be used for final grading. This field specifies\n\
                    \            how the ultimate submission should be determined."
                  nullable: false
                  readOnly: false
                  allOf:
                  - $ref: '#/components/schemas/UltimateSubmissionPolicy'
                hide_ultimate_submission_fdbk:
                  description: "A hard override that indicates that ultimate\n   \
                    \         submission feedback should not be shown, even if the\n\
                    \            appropriate criteria are met."
                  nullable: false
                  readOnly: false
                  type: boolean
                instructor_files:
                  description: ''
                  nullable: false
                  readOnly: true
                  type: array
                  items:
                    $ref: '#/components/schemas/InstructorFile'
                expected_student_files:
                  description: ''
                  nullable: false
                  readOnly: true
                  type: array
                  items:
                    $ref: '#/components/schemas/ExpectedStudentFile'
                has_handgrading_rubric:
                  description: Whether this project has a handgrading rubric.
                  readOnly: true
                  type: boolean
      tags: *id015
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        pk: *id047
        name:
          description: "The name used to identify this project.\n            Must\
            \ be non-empty and non-null.\n            Must be unique among Projects\
            \ associated with\n            a given course.\n            This field\
            \ is REQUIRED."
          nullable: false
          type: string
        last_modified:
          description: ''
          nullable: false
          type: string
          format: date-time
        course:
          description: "The Course this project belongs to.\n            This field\
            \ is REQUIRED."
          nullable: false
          type: integer
          format: id
        visible_to_students:
          description: "Whether information about this Project can\n            be\
            \ viewed by students."
          nullable: false
          type: boolean
        closing_time:
          description: "The date and time that this project should stop\n        \
            \    accepting submissions.\n            A value of None indicates that\
            \ this project should\n            stay open.\n            In the API,\
            \ this field is hidden from non-admins."
          nullable: true
          type: string
          format: date-time
        soft_closing_time:
          description: "The date and time that should be displayed as the\n      \
            \      due date for this project. Unlike closing_time,\n            soft_closing_time\
            \ does not affect whether submissions are\n            actually accepted.\n\
            \            If not None and closing_time is not None, this value must\
            \ be\n            less than (before) closing_time."
          nullable: true
          type: string
          format: date-time
        disallow_student_submissions:
          description: "A hard override that indicates that students should\n    \
            \        be prevented from submitting even if visible_to_students is\n\
            \            True and it is before closing_time."
          nullable: false
          type: boolean
        disallow_group_registration:
          description: "A hard override that indicates that students should\n    \
            \        not be able to send, accept, or reject group\n            invitations."
          nullable: false
          type: boolean
        guests_can_submit:
          description: "By default, only admins, staff, and students\n           \
            \ for a given Course can view and submit to its Projects.\n          \
            \  When True, submissions will be accepted from guests\n            with\
            \ the following caveats:\n                - Guests must be given a direct\
            \ link to the project.\n                - When group work is allowed,\
            \ guests can\n                only be in groups with other guests."
          nullable: false
          type: boolean
        min_group_size:
          description: "The minimum number of students that can work in a\n      \
            \      group on this project.\n            Must be >= 1.\n           \
            \ Must be <= max_group_size."
          nullable: false
          type: integer
        max_group_size:
          description: "The maximum number of students that can work in a\n      \
            \      group on this project.\n            Must be >= 1.\n           \
            \ Must be >= min_group_size."
          nullable: false
          type: integer
        submission_limit_per_day:
          description: "The number of submissions each group is allowed per\n    \
            \        day before either reducing feedback or preventing further\n \
            \           submissions. A value of None indicates no limit."
          nullable: true
          type: integer
        allow_submissions_past_limit:
          description: "Whether to allow additional submissions after a\n        \
            \    group has submitted submission_limit_per_day times."
          nullable: false
          type: boolean
        groups_combine_daily_submissions:
          description: "If True, group members can \"pool\" their daily submissions.\n\
            \            For example, if submission_limit_per_day is 3,\n        \
            \    a group with 2 members would get 6 submissions per day."
          nullable: false
          type: boolean
        submission_limit_reset_time:
          description: "The time that marks the beginning and end of the 24\n    \
            \        hour period during which submissions should be counted\n    \
            \        towards the daily limit. Defaults to 00:00:00."
          nullable: false
          type: string
          format: time
        submission_limit_reset_timezone:
          description: "The timezone to use when computing how many\n            submissions\
            \ a group has made in a 24 hour period."
          nullable: false
          type: string
          format: timezone
        num_bonus_submissions:
          description: ''
          nullable: false
          type: integer
        total_submission_limit:
          description: "The maximum number of times a Group can submit to\n      \
            \      this Project EVER."
          nullable: true
          type: integer
        allow_late_days:
          description: "Whether to allow the use of late days for submitting\n   \
            \         past the deadline."
          nullable: false
          type: boolean
        ultimate_submission_policy:
          description: "The \"ultimate\" submission for a group is the one\n     \
            \       that will be used for final grading. This field specifies\n  \
            \          how the ultimate submission should be determined."
          nullable: false
          allOf:
          - $ref: '#/components/schemas/UltimateSubmissionPolicy'
        hide_ultimate_submission_fdbk:
          description: "A hard override that indicates that ultimate\n           \
            \ submission feedback should not be shown, even if the\n            appropriate\
            \ criteria are met."
          nullable: false
          type: boolean
        instructor_files:
          description: ''
          nullable: false
          type: array
          items:
            $ref: '#/components/schemas/InstructorFile'
        expected_student_files:
          description: ''
          nullable: false
          type: array
          items:
            $ref: '#/components/schemas/ExpectedStudentFile'
        has_handgrading_rubric:
          description: Whether this project has a handgrading rubric.
          type: boolean
    UltimateSubmissionPolicy:
      type: string
      enum:
      - most_recent
      - best_basic_score
      - best
    ExpectedStudentFile:
      type: object
      properties:
        pk: *id047
        project:
          description: ''
          nullable: false
          type: integer
          format: id
        pattern:
          description: "A shell-style file pattern suitable for\n            use with\
            \ Python's fnmatch.fnmatch()\n            function (https://docs.python.org/3.5/library/fnmatch.html)\n\
            \            This string must be a legal UNIX filename and may not be\n\
            \            '..' or '.'.\n            NOTE: Patterns for a given project\
            \ must not overlap,\n                otherwise the behavior is undefined."
          nullable: false
          type: string
        min_num_matches:
          description: "The minimum number of submitted student files that\n     \
            \       should match the pattern. Must be non-negative."
          nullable: false
          type: integer
        max_num_matches:
          description: "The maximum number of submitted student files that\n     \
            \       can match the pattern. Must be >= min_num_matches"
          nullable: false
          type: integer
        last_modified:
          description: ''
          nullable: false
          type: string
          format: date-time
    InstructorFile:
      type: object
      properties:
        pk: *id047
        project:
          description: ''
          nullable: false
          type: integer
          format: id
        name:
          description: ''
          nullable: false
          type: string
        last_modified:
          description: ''
          nullable: false
          type: string
          format: date-time
        size:
          description: ''
          type: integer
...